Tests:Twisted:gabbletest generalize and modernize authenticator
 * Move TlsAuthenticator from tls to gabbletest
 * Move features dict from XmppAuthenticator.streamIQ to class attrbiute
 * Handle errors to improve failure reporting
